requirement.txt
pycodestyle==2.6.0
numpy==1.19.4
opencv-python==4.2.0.32
h5py==3.1.0
pandas==1.1.4
Pillow==8.0.1
pycocotools==pip 2.0.2